Added fallback for doc utils web worker. Optimisez performant local strage
